requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113 
torchvision==0.12.0+cu113
numpy
tqdm
matplotlib
imageio
plotly_express